Inventory functionality complete
Inventory is fully functional, items appear in the inventory correctly. Might need some more visual tweaks but overall looks good. Last thing to do is add interface for adding items to the inventory as right now it just handles starting items.

Also did some black magic in the UI builder to make it scale correctly so the inventory should look good no matter the resolution. Did this by setting the grid to have a width and height of 100% and have auto margins. Somehow that works and makes it scale without breaking aspect ratios of slots but definitely something to keep an eye on for problems.