Reworked placement
Redid placement to make icons children of the slot elements as there seemed to be some weird inaccuracies possibly from floating point numbers being converted into ints by the UI which made it impossible to manually set the size of the icons. So instead of manually setting the size I made the icons children of the slots so they would just expand to take up the whole slot space automatically. The only thing this breaks is items bigger then 1x1 but since we're not too worried about those for now that's an issue to worry about later.