End of day ui functional
End of day ui is in and functional now. Just drag the end of day prefab into the main level and it should be able to get the data from the player data script on the player prefab on its own. To end the day just call the public end day function on the end day code.