Fixed bug, added starting rotation direction, and fully added to menu
Fixed a bug that stopped the button animation from working with multiple buttons. Added a starting direction variable so you can control what direction a button starts rotated, lets you alternate them starting off which looks nicer, and added it to the exit button in the main menu as well.